Create Exercise 2 Action Node
Exercise 2: Action Node
Instructions
You can use this exercise to complete the implementation of the ActionNode Class. First copy and paste the Node Class that you have completed in Exercise 1.
Now complete the implementation of the ActionNode class, as described in Exercise 2 of the coursework specification.
How to Compile and Run Your Code
You can compile and run your code using the “RUN IT” button below. This will compile your java code into byte code, then run the YourTestCode file.
Run it
Explanation of Automated Tests
There are two automated tests.
test1: Is the same as the test1 from the previous exercise, except that we create ActionNode objects, and assign them to a variable of type Node.
test2: Creates an ActionNode object, checks that its getNext method initially returns null. Then creates a second ActionNode object, sets this as the next node for the first (using the setNext method), then checks that the getNext method, when called on the first node, returns a reference to the second.